+++ projectroles/__init__.py
@@ -2,15 +2,10 @@
 SODAR project and role management
 """
 
-from ._version import get_versions
+from . import _version  # noqa
 
-__version__ = get_versions()['version']
-del get_versions
+__version__ = _version.get_versions()['version']
 
 default_app_config = (
     'projectroles.apps.ProjectrolesConfig'  # pylint: disable=invalid-name
 )
-
-from . import _version  # noqa
-
-__version__ = _version.get_versions()['version']